chuyen double thanh float, class animate add SpriteBatch

+++ core/src/com/mygdx/game/entities/DynamicEntity/Bomber.java
@@ -3,7 +3,7 @@ package com.mygdx.game.entities.DynamicEntity;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
 import com.mygdx.game.entities.AnimatedEntity;
-import com.mygdx.game.entities.gamesys.GameManager;
+import com.mygdx.game.gamesys.GameManager;
 
 public class Bomber extends AnimatedEntity {
     private int code = 0; //Mã phím vừa bấm.
@@ -15,7 +15,7 @@ public class Bomber extends AnimatedEntity {
         code = Input.Keys.S;
     }
 
-    public Bomber(double x, double y) {
+    public Bomber(float x, float y) {
         super(x, y);
         textureAtlas = GameManager.playerDownStatic.getKey();
         animation = GameManager.playerDownStatic.getValue();
